app/requirements.txt
fastapi
uvicorn
psycopg2-binary
redis
spacy
numpy
python-dotenv
openai
cryptography